Fixed handling of nmol = 0 in esolan.
The function energy calls esolan(0) to get the solvent energy if itysol > 0.
Esolan should calculate the solvent energy using all atoms in the system in
this case. I added an if statement at the beginning of esolan() that takes
care of this now. Further rewrote energy() so that it assigns the result
returned by esolan() to teysl, which in turn will be assigned to eysl in the
end.