Updated version of Lund force field functions
The Lund force field functions were debugged. A new library file
lib.lun was added to describe the geometrical parameters of
protein molecules consistent with the force field. With this
library file defining the bond lengths and bond angles, energy
values calculated with SMMP with the Lund force field can be
directly compared with those calculated with PROFASI.