Fixing omega when structure is read from a PDB file
For Lund force field, the fixing of the omega angle did not
work when a structure was read in from a PDB file. This is now
fixed.

Also, fake nh2g, nh2 and cooh groups were added to lib.lun so
that no matter what is requested, if lib.lun is used, the
end groups are charged.