Redirected standard out to logString.
SMMP produced a lot of log messages.  This became an issue when run in massively
parallel environments.  I replaced all writes to standard out to a write to logString.
The next step is to pass this string to a function that writes the messages to a log
file according to their importance and the chosen log level.